requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-json-report>=1.5.0
google-generativeai>=0.3.0 